refactor(core): Modularize SimulationEngine and refactor UI components
This commit introduces a major architectural refactoring to improve separation of concerns, testability, and overall code organization. The monolithic `SimulationEngine` has been decomposed into smaller, single-responsibility managers, and several UI components have been extracted for better reusability.

The `SimulationEngine`'s role has been refined to be a high-level orchestrator. Most of its internal logic has been extracted into the following specialized modules:

-   **`PopulationManager`**: Encapsulates all logic for dynamic ecosystem balancing. It monitors population trends and decides when to spawn new actors like birds, eagles, or herbicide planes.
-   **`AsyncFlowerFactory`**: Manages all asynchronous communication with the `flower.worker.ts`. It decouples expensive WASM genetics calls from the main simulation loop by handling the flower request queue and `FlowerSeed` placeholder logic.
-   **`EcosystemManager`**: A new home for global, system-wide behaviors that don't belong to a single actor, such as nutrient healing and insect reproduction.
-   **`environmentManager`**: Contains the logic for updating the simulation's season, temperature, humidity, and weather events each tick.

Several UI components have been extracted to promote reusability and simplify parent components:

-   Created `CollapsibleSection` from logic previously in `Controls`.
-   Created `ConfirmationModal` and `SeedBankCard` from logic in `SeedBankPanel`.
-   Created `EnvironmentDisplay` to separate environment UI from `App.tsx`.
-   Added a dedicated close button to the `FlowerDetailsPanel` for better UX and simpler E2E tests.

The monolithic `src/types.ts` file has been split into a structured `src/types/` directory with clear categories (`actors`, `base`, `ui`, `services`), improving type navigation and organization.

Documentation in `README.md` and `Planning.md` has been updated to reflect this new modular architecture.
