+++ src/components/SeedBankPanel.tsx
@@ -1,78 +1,12 @@
 import React, { useState, useEffect, useCallback, useRef } from 'react';
 import { db } from '../services/db';
 import type { SeedBankEntry } from '../types';
-import { LoaderIcon, Trash2Icon, DownloadIcon, View3DIcon } from './icons';
+import { LoaderIcon, Trash2Icon } from './icons';
 import { Modal } from './Modal';
 import { Flower3DViewer } from './Flower3DViewer';
 import { flowerService } from '../services/flowerService';
-
-const ConfirmationModal: React.FC<{ onConfirm: () => void; onCancel: () => void; }> = ({ onConfirm, onCancel }) => (
-    <div className="fixed inset-0 bg-black/70 backdrop-blur-sm flex items-center justify-center z-50 p-4" onClick={onCancel}>
-        <div className="bg-surface border border-border rounded-lg shadow-xl w-full max-w-md p-6" onClick={(e) => e.stopPropagation()}>
-            <h3 className="text-xl font-bold text-primary-light">Clear Seed Bank?</h3>
-            <p className="text-secondary my-4">This will remove all saved champion flowers and cannot be undone. New flowers will be completely random until new champions are saved.</p>
-            <div className="flex justify-end space-x-4">
-                <button onClick={onCancel} className="px-4 py-2 bg-surface-hover hover:bg-border/20 text-primary-light font-semibold rounded-md transition-colors">Cancel</button>
-                <button onClick={onConfirm} className="px-4 py-2 bg-accent-red/80 hover:bg-accent-red text-white font-semibold rounded-md transition-colors">Confirm & Clear</button>
-            </div>
-        </div>
-    </div>
-);
-
-interface SeedBankCardProps {
-    entry: SeedBankEntry;
-    onDownload: (genome: string, id: string) => void;
-    onView3D: (genome: string, sex: 'male' | 'female' | 'both') => void;
-}
-
-const SeedBankCard: React.FC<SeedBankCardProps> = ({ entry, onDownload, onView3D }) => {
-    let title = '';
-    let value = '';
-
-    switch (entry.category) {
-        case 'longestLived':
-            title = 'Longest Lived';
-            value = `${entry.value} ticks`;
-            break;
-        case 'mostToxic':
-            title = 'Most Toxic';
-            value = `${(entry.value * 100).toFixed(0)}% Toxicity`;
-            break;
-        case 'mostHealing':
-            title = 'Most Healing';
-            // Since healing is negative, multiply by -1 for display
-            value = `${(entry.value * -100).toFixed(0)}% Healing`;
-            break;
-    }
-    
-    return (
-        <div className="bg-surface-hover/50 rounded-lg p-3 flex flex-col">
-            <h4 className="font-semibold text-primary">{title}</h4>
-            <p className="text-sm text-accent-yellow font-bold mb-2">{value}</p>
-            <div className="bg-black/50 rounded flex items-center justify-center p-2 aspect-square">
-                 <img src={entry.imageData} alt={`Champion flower for ${title}`} className="w-full h-full object-contain"/>
-            </div>
-            <div className="grid grid-cols-2 gap-2 mt-3 pt-3 border-t border-border/50">
-                <button 
-                    onClick={() => onDownload(entry.genome, entry.category)} 
-                    className="flex items-center justify-center gap-2 text-sm p-2 bg-surface-hover rounded-md hover:bg-border/20 transition-colors"
-                    title="Download genome as a JSON file"
-                >
-                    <DownloadIcon className="w-4 h-4" />
-                    Download
-                </button>
-                 <button 
-                    onClick={() => onView3D(entry.genome, entry.sex || 'both')} // Fallback for old saves
-                    className="flex items-center justify-center gap-2 text-sm p-2 bg-surface-hover rounded-md hover:bg-border/20 transition-colors"
-                    title="View 3D model"
-                >
-                    <View3DIcon className="w-4 h-4" />
-                    View 3D
-                </button>
-            </div>
-        </div>
-    );
-};
+import { ConfirmationModal } from './ConfirmationModal';
+import { SeedBankCard } from './SeedBankCard';
 
 interface SeedBankPanelProps {
     isRunning: boolean;
@@ -182,7 +116,12 @@ export const SeedBankPanel: React.FC<SeedBankPanelProps> = ({ isRunning, setIsRu
                 </div>
             )}
 
-            {showConfirmModal && <ConfirmationModal onConfirm={handleClearBank} onCancel={() => setShowConfirmModal(false)} />}
+            {showConfirmModal && <ConfirmationModal 
+                onConfirm={handleClearBank} 
+                onCancel={() => setShowConfirmModal(false)}
+                title="Clear Seed Bank?"
+                message="This will remove all saved champion flowers and cannot be undone. New flowers will be completely random until new champions are saved."
+            />}
             
             <Modal
                 isOpen={is3DViewerOpen}
