--- src/types.ts
@@ -1,342 +1,0 @@
-export type WindDirection = 'N' | 'NE' | 'E' | 'SE' | 'S' | 'SW' | 'W' | 'NW';
-
-export type NotificationMode = 'toasts' | 'log' | 'both';
-
-export type Season = 'Spring' | 'Summer' | 'Autumn' | 'Winter';
-export type WeatherEventType = 'heatwave' | 'coldsnap' | 'heavyrain' | 'drought' | 'none';
-
-export interface WeatherEvent {
-    type: WeatherEventType;
-    duration: number; // Ticks remaining
-}
-
-export interface EnvironmentState {
-    currentTemperature: number;
-    currentHumidity: number;
-    season: Season;
-    currentWeatherEvent: WeatherEvent;
-}
-export interface SimulationParams {
-    gridWidth: number;
-    gridHeight: number;
-    initialFlowers: number;
-    initialInsects: number;
-    initialBirds: number;
-    humidity: number;
-    temperature: number;
-    windDirection: WindDirection;
-    windStrength: number;
-    flowerDetailRadius: number;
-    herbicideFlowerDensityThreshold: number;
-    herbicideDamage: number;
-    herbicideSmokeLifespan: number;
-    herbicideCooldown: number;
-    herbicideSmokeExpansionCount: number;
-    notificationMode: NotificationMode;
-    // Seasonal Cycle Parameters
-    seasonLengthInTicks: number;
-    temperatureAmplitude: number;
-    humidityAmplitude: number;
-    // Weather Event Parameters
-    weatherEventChance: number;
-    heatwaveTempIncrease: number;
-    coldsnapTempDecrease: number;
-    heavyRainHumidityIncrease: number;
-    droughtHumidityDecrease: number;
-    weatherEventMinDuration: number;
-    weatherEventMaxDuration: number;
-}
-
-export interface Coord {
-    x: number;
-    y: number;
-}
-
-export interface Actor extends Coord {
-    id: string;
-}
-
-// Raw genetic effects from WASM
-export interface FlowerEffects {
-    vitality: number;
-    agility: number;
-    intelligence: number;
-    strength: number;
-    luck: number;
-}
-
-// The raw JSON object returned by the WASM getFlowerStats function
-export interface FlowerGenomeStats {
-    health: number; // This is maxHealth
-    stamina: number; // This is maxStamina
-    minTemperature: number;
-    maxTemperature: number;
-    maturationPeriod: number;
-    sex: 'male' | 'female' | 'both';
-    toxicityRate: number;
-    effects: FlowerEffects;
-}
-
-// The processed stats used within the simulation
-export interface FlowerStats {
-    maxHealth: number;
-    maxStamina: number;
-    nutrientEfficiency: number;
-    minTemperature: number;
-    maxTemperature: number;
-    maturationPeriod: number;
-    sex: 'male' | 'female' | 'both';
-    toxicityRate: number;
-    effects: FlowerEffects;
-}
-
-export interface Flower extends Actor, FlowerStats {
-    type: 'flower';
-    genome: string;
-    imageData: string;
-    health: number; // Current health
-    stamina: number; // Current stamina
-    age: number;
-    isMature: boolean;
-}
-
-export interface FlowerSeed extends Actor {
-    type: 'flowerSeed';
-    imageData: string;
-    health: number;
-    maxHealth: number;
-    age: number;
-}
-
-export interface Insect extends Actor {
-    type: 'insect';
-    pollen: {
-        genome: string;
-        sourceFlowerId: string;
-    } | null; // Genome and source ID of last visited flower
-    emoji: string;
-    lifespan: number;
-    reproductionCooldown?: number;
-}
-
-export interface Bird extends Actor {
-    type: 'bird';
-    target: {x: number, y: number} | null; // target insect coordinates
-    patrolTarget: {x: number, y: number} | null;
-}
-
-export interface Eagle extends Actor {
-    type: 'eagle';
-    target: {x: number, y: number} | null; // target bird coordinates
-}
-
-export interface Nutrient extends Actor {
-    type: 'nutrient';
-    lifespan: number;
-}
-
-export interface Egg extends Actor {
-    type: 'egg';
-    hatchTimer: number;
-    insectEmoji: string;
-}
-
-export interface HerbicidePlane extends Actor {
-    type: 'herbicidePlane';
-    dx: number; // Current direction x
-    dy: number; // Current direction y
-    turnDx: number; // Direction to turn x
-    turnDy: number; // Direction to turn y
-    stride: number;
-}
-
-export interface HerbicideSmoke extends Actor {
-    type: 'herbicideSmoke';
-    lifespan: number;
-    canBeExpanded: number;
-}
-
-export type CellContent = Flower | Insect | Bird | Nutrient | Egg | Eagle | HerbicidePlane | HerbicideSmoke | FlowerSeed;
-
-export type Grid = (CellContent[])[][];
-
-export type PopulationTrend = 'growing' | 'declining' | 'stable';
-
-export interface AppEvent {
-  message: string;
-  type: 'info' | 'success' | 'error';
-  importance: 'low' | 'high';
-  tick?: number;
-  timestamp?: number;
-}
-
-export interface LogEntry extends AppEvent {
-  id: string;
-}
-
-export interface ToastMessage {
-  id: string;
-  message: string;
-  type: 'info' | 'success' | 'error';
-  count?: number;
-  key?: number;
-}
-
-export type ActorAddDelta = {
-    type: 'add';
-    actor: CellContent;
-};
-export type ActorUpdateDelta = {
-    type: 'update';
-    id: string;
-    changes: Partial<CellContent>;
-};
-export type ActorRemoveDelta = {
-    type: 'remove';
-    id: string;
-};
-export type ActorDelta = ActorAddDelta | ActorUpdateDelta | ActorRemoveDelta;
-
-
-export interface FEService {
-    initialize(): Promise<void>;
-    setParams(params: FEParams): void;
-    getParams(): FEParams;
-    makeFlower(): Promise<{genome: string, image: string}>;
-    makePetals(): Promise<{ genome: string; image: string }>;
-    makePetalLayer(layer: number): Promise<{ genome: string; image: string }>;
-    makeStem(): Promise<{ genome: string; image: string }>;
-    reproduce(father: string, mother: string): Promise<{genome: string, image: string}>;
-    mutate(original: string, addNodeRate?: number, addConnRate?: number, removeConnRate?: number, perturbWeightsRate?: number, enableRate?: number, disableRate?: number, actTypeRate?: number): Promise<{genome: string, image: string}>;
-    getFlowerStats(genome: string, humidity?: number, temperature?: number, altitude?: number, terrainType?: number): Promise<FlowerGenomeStats>;
-    drawFlower(genome: string): Promise<{genome: string, image: string}>;
-    drawPetals(genome: string): Promise<{ genome: string; image: string }>;
-    drawPetalLayer(genome: string, layer: number): Promise<{ genome: string; image: string }>;
-    draw3DFlower(genome: string, sex: 'male' | 'female' | 'both'): Promise<string>;
-    drawEmissive3DFlower(genome: string, sex: 'male' | 'female' | 'both'): Promise<string>;
-}
-
-export interface FEParams {
-    radius: number;
-    numLayers: number;
-    P: number;
-    bias: number;
-}
-
-// --- Challenges & Analytics ---
-
-export type ChallengeId = 'survival-1' | 'survival-2' | 'survival-3' | 'predation-1' | 'predation-2' | 'predation-3' |
-'circle-of-life' | 'pest-control' | 'bountiful-harvest' | 'poison-garden' | 'peak-performer' | 'the-swarm-1' | 'the-swarm-2' | 'avian-sanctuary-1' | 'avian-sanctuary-2' | 'unchecked-growth';
-
-export interface Challenge {
-    id: ChallengeId;
-    title: string;
-    description: string;
-    goal: number;
-    progress: number;
-    completed: boolean;
-    metric: keyof TickSummary;
-    aggregator: 'max' | 'sum';
-}
-
-export interface ChallengeState {
-    challenges: Challenge[];
-    processTick: (summary: TickSummary) => void;
-}
-
-export interface AnalyticsDataPoint {
-    tick: number;
-    flowers: number;
-    insects: number;
-    birds: number;
-    eagles: number;
-    eggCount: number;
-    herbicidePlanes: number;
-    herbicideSmokes: number;
-    reproductions: number;
-    insectsEaten: number;
-    eggsEaten: number;
-    insectsDiedOfOldAge: number;
-    eggsLaid: number;
-    insectsBorn: number;
-    totalBirdsHunted: number;
-    totalHerbicidePlanesSpawned: number;
-    nutrientCount: number;
-    flowerDensity: number;
-    avgHealth: number;
-    maxHealth: number;
-    maxToxicity: number;
-    avgStamina: number;
-    maxStamina: number;
-    avgNutrientEfficiency: number;
-    avgMaturationPeriod: number;
-    avgVitality: number;
-    avgAgility: number;
-    avgStrength: number;
-    avgIntelligence: number;
-    avgLuck: number;
-    tickTimeMs: number;
-    renderTimeMs: number;
-    currentTemperature?: number;
-    currentHumidity?: number;
-    season?: Season;
-    weatherEvent?: WeatherEventType;
-}
-
-export interface AnalyticsState {
-    history: AnalyticsDataPoint[];
-    addDataPoint: (data: { summary: TickSummary; renderTimeMs: number }) => void;
-    reset: () => void;
-}
-
-export interface TickSummary {
-    tick: number;
-    flowerCount: number;
-    insectCount: number;
-    birdCount: number;
-    eagleCount: number;
-    eggCount: number;
-    herbicidePlaneCount: number;
-    herbicideSmokeCount: number;
-    reproductions: number;
-    insectsEaten: number; // In this tick
-    totalInsectsEaten: number; // Cumulative
-    totalBirdsHunted: number; // Cumulative
-    totalHerbicidePlanesSpawned: number; // Cumulative
-    nutrientCount: number; // Current tick
-    flowerDensity: number; // Current tick
-    maxFlowerAge: number;
-    eggsLaid: number;
-    insectsBorn: number;
-    eggsEaten: number;
-    insectsDiedOfOldAge: number;
-    avgHealth: number;
-    avgStamina: number;
-    maxHealth: number;
-    maxToxicity: number;
-    maxStamina: number;
-    avgNutrientEfficiency: number;
-    avgMaturationPeriod: number;
-    avgVitality: number;
-    avgAgility: number;
-    avgStrength: number;
-    avgIntelligence: number;
-    avgLuck: number;
-    tickTimeMs: number;
-    currentTemperature: number;
-    currentHumidity: number;
-    season: Season;
-    weatherEvent: WeatherEventType;
-}
-
-// --- Seed Bank ---
-
-export type SeedBankCategory = 'longestLived' | 'mostToxic' | 'mostHealing';
-
-export interface SeedBankEntry {
-    category: SeedBankCategory;
-    genome: string;
-    value: number;
-    imageData: string;
-    sex: 'male' | 'female' | 'both';
-}
